feat: Add method to retrieve catalog areas by geometry
This commit adds a new method `getCatalogAreasByGeometry` to the `CadastralParcel` model. This method retrieves an array of catalog areas that intersect with the cadastral parcel's geometry. It takes a catalog ID as a parameter and returns an array of catalog area IDs.

The `CadastralParcel` Nova resource has been updated to display the retrieved catalog areas in the detail view using a new field called "Catalog Areas". The field dynamically generates HTML links to each catalog area's detail page.

Additionally, some commented-out code related to calculating the total estimated value of cadastral parcels for owners has been temporarily disabled in the `Owner` Nova resource.