feat: Add migration to enable postgis_raster extension
This commit adds a new migration file that enables the "postgis_raster" extension in the database. The migration uses the `CREATE EXTENSION IF NOT EXISTS` statement to ensure that the extension is only created if it doesn't already exist. Additionally, a corresponding `down` method is implemented to drop the extension if necessary.