feat: Add hiking_routes_min_dist column to catalog_areas table
This commit adds a new migration file that adds the `hiking_routes_min_dist` column to the `catalog_areas` table. The column is of type float and has a default value of 0.0.

Additionally, in the `EnrichCatalogAreaWithHikingRouteStatsCommand.php` file, if there are no hiking route stats available for an area, the command now sets the `hiking_routes_min_dist` attribute of the area using the `getHikingRouteMinDist()` method before saving it.

These changes ensure that we have accurate information about hiking routes in each catalog area.