feat: Add command to enrich catalog areas with street info
This commit adds a new command `sisteco:enrich-streets` that adds street information to catalog areas. The command retrieves the minimum distance between each catalog area and its corresponding streets using a SQL query. The result is then saved in the `streets_min_dist` column of the `catalog_areas` table.

The new command is executed in the deployment script for release 1.1.0, ensuring that street information is enriched during the deployment process.

Semantic versioning guidelines have been followed for this commit message.